+++ tests/data/api_mock_responses/login_with_email_and_password.py
@@ -1,0 +1,25 @@
+login_with_email_and_password_success_headers = {
+    "content-type": "application/json; charset=utf-8",
+    "set-cookie": "access-token=eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ1c2VyMTIzIiwiaWF0IjoxNzI1MDA0MzI1LCJpc3"
+                  "MiOiJNeUFQSSJ9.Xy5uEfgD7bBq5I7Xy8VhUzvB8JOpYVzphEKvAFTnXDbzHlAXP9xFi84gFkcT3_kKPfbbd97xHyEhQAYypbJtJ"
+                  "xIcEAWtVdJhxzSbcAd7dR9SKXzC6X_S1lpg8Z9QjzATQBRlhEM7Ru1gXt2AXZxn8pq5GswWws9BuL4OFEf8joI5AFH11_wO_NUpO"
+                  "yo1JoZxEU0g5xnFGAtg9HiLArkeK1e7E0d2aJfqX5OOGPCg-wZLzTWShpCwVdSyzEFvB1RuJCL_N4Whw2A0LvF8n1HRR6gfh1zIh"
+                  "Ye3LVJBUOm29W8f-KzNZGz_9Kc2tXyGk-4hd47UzBaG6y42AzmcKbQeCl1M0V7N4cTkD5XFE34Z11wto-jlAVk4d9Z-N0jf21wGS"
+                  "ljHh1QuAZKi1LfX1gKyxZPEK9Ir2XhmA3gWBS5TlXPhASot15AOSNntty0RAnD4zFOCtwX_N69Wq_JR0V4WEt79fhhlMB9ar5FPW"
+                  "WgTf2IqaAWlVfX8RWTpTV-vvWErfLqkXmdde-P0iDEm3C2x-d3NF5A8Amf7L0MeuyJ_4uYRaUq8pZyJ_hXidDSKvL3_Zshk7rTME"
+                  "6mdUir6SYdSLfgwN98NGRY6IcXFEugL25L-j0NR1FnV0S_eVHw2z8OcHZZTDqF_XgxFCuxNLLnN5FDJwM1uDqFFPViIE2y24wDVe"
+                  "A; Path=/; Domain=.dehancer.com; Max-Age=31536000; HttpOnly; Secure, "
+                  "auth=c54a1a45-3b36-4c59-9a11-8c0ed8f0a66f; Path=/; "
+                  "Domain=.dehancer.com; Max-Age=2678400; HttpOnly; Secure",
+    "strict-transport-security": "max-age=31536000",
+}
+
+login_with_email_and_password_success_response = {
+    "success": True,
+}
+
+login_with_email_and_password_not_success_response = {
+    "success": False,
+}
+
+login_with_email_and_password_invalid_response = "Error"  # noqa: S105